backend/requirements.txt
asgiref==3.6.0
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
Django==3.2
django-cors-headers==4.0.0
django-extra-fields==3.0.2
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-commas==2.1.0
flake8-isort==6.0.0
flake8-plugin-utils==1.3.2
flake8-quotes==3.3.2
flake8-return==1.2.0
gunicorn==20.1.0
idna==3.4
isort==5.11.5
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
pep8-naming==0.13.3
Pillow==9.5.0
platformdirs==3.5.1
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.7.0
PyPDF2==2.12.1
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
uritemplate==4.1.1
urllib3==2.0.2
webcolors==1.13
